+++ NEWS.md
@@ -6,6 +6,16 @@ This file documents the development of the package as well as open issues or poi
 
 
 
+### Version in 0.9.15
+Date: 2018.09.18
+
+Changes:
+
+*   plotDensity function added
+*   attributes added to data.frame to increase usability
+
+
+
 ### Version in 0.9.13
 Date: 2018.09.16